refactor(action-collection): context-based dialog and component extraction
