+++ src/entrypoints/options/app.tsx
@@ -2,7 +2,6 @@ import type { ComponentType } from "react"
 import { lazy, Suspense } from "react"
 import { Route, Routes } from "react-router"
 import { ROUTE_DEFS } from "./app-sidebar/nav-items"
-import { ActionCollectionInstallListener } from "./components/action-collection-install-listener"
 import { GeneralPage } from "./pages/general"
 
 type RoutePath = (typeof ROUTE_DEFS)[number]["path"]
@@ -44,16 +43,13 @@ function RouteLoadingFallback() {
 
 export default function App() {
   return (
-    <>
-      <ActionCollectionInstallListener />
-      <Suspense fallback={<RouteLoadingFallback />}>
-        <Routes>
-          {ROUTE_DEFS.map(({ path }) => {
-            const Component = ROUTE_COMPONENTS[path]
-            return <Route key={path} path={path} element={<Component />} />
-          })}
-        </Routes>
-      </Suspense>
-    </>
+    <Suspense fallback={<RouteLoadingFallback />}>
+      <Routes>
+        {ROUTE_DEFS.map(({ path }) => {
+          const Component = ROUTE_COMPONENTS[path]
+          return <Route key={path} path={path} element={<Component />} />
+        })}
+      </Routes>
+    </Suspense>
   )
 }
