+++ src/entrypoints/options/pages/custom-actions/components/featured-collections.tsx
@@ -1,13 +1,13 @@
+import type { ActionCollectionStoreListItem } from "@read-frog/api-contract"
 import { Icon } from "@iconify/react"
-import { useState } from "react"
 import { i18n } from "#imports"
 import { Skeleton } from "@/components/ui/base-ui/skeleton"
 import { useActionCollectionPublicCollections } from "@/orpc/action-collection"
-import { ActionCollectionDetailDialog } from "./ai-feature-store"
+import { useActionCollectionDetailDialog } from "./action-collection-detail-dialog"
 
 export function FeaturedActionCollections() {
   const { items, isLoading } = useActionCollectionPublicCollections({ limit: 6, sort: "all_time" })
-  const [detailId, setDetailId] = useState<number | null>(null)
+  const { openDetail } = useActionCollectionDetailDialog()
 
   if (!isLoading && items.length === 0) {
     return null
@@ -24,37 +24,32 @@ export function FeaturedActionCollections() {
               <Skeleton key={key} className="h-[76px] w-52 shrink-0 rounded-xl" />
             ))
           : items.map(item => (
-              <button
-                key={item.collectionId}
-                type="button"
-                className="group flex w-52 shrink-0 cursor-pointer flex-col gap-1 rounded-xl border bg-card p-3 text-left transition-colors hover:border-primary/40 hover:bg-accent/40"
-                onClick={() => setDetailId(item.collectionId)}
-              >
-                <div className="flex items-center gap-1.5">
-                  <Icon icon={item.icon ?? "tabler:sparkles"} className="size-4 shrink-0 text-foreground/80" />
-                  <span className="truncate text-sm font-semibold text-foreground">{item.name}</span>
-                </div>
-                <div className="truncate text-xs text-muted-foreground">{item.description}</div>
-                <div className="flex items-center justify-between gap-2 text-xs text-muted-foreground/70">
-                  <span className="truncate text-muted-foreground">{item.author.displayName}</span>
-                  <span className="flex shrink-0 items-center gap-1">
-                    <Icon icon="tabler:download" className="size-3.5" />
-                    {item.installCount}
-                  </span>
-                </div>
-              </button>
+              <FeaturedCard key={item.collectionId} item={item} onOpen={() => openDetail(item.collectionId)} />
             ))}
       </div>
-      <ActionCollectionDetailDialog
-        open={detailId != null}
-        collectionId={detailId}
-        onOpenChange={(open) => {
-          if (!open) {
-            setDetailId(null)
-          }
-        }}
-        onInstalled={() => setDetailId(null)}
-      />
     </div>
+  )
+}
+
+function FeaturedCard({ item, onOpen }: { item: ActionCollectionStoreListItem, onOpen: () => void }) {
+  return (
+    <button
+      type="button"
+      className="group flex w-52 shrink-0 cursor-pointer flex-col gap-1 rounded-xl border bg-card p-3 text-left transition-colors hover:border-primary/40 hover:bg-accent/40"
+      onClick={onOpen}
+    >
+      <div className="flex items-center gap-1.5">
+        <Icon icon={item.icon ?? "tabler:sparkles"} className="size-4 shrink-0 text-foreground/80" />
+        <span className="truncate text-sm font-semibold text-foreground">{item.name}</span>
+      </div>
+      <div className="truncate text-xs text-muted-foreground">{item.description}</div>
+      <div className="flex items-center justify-between gap-2 text-xs text-muted-foreground/70">
+        <span className="truncate text-muted-foreground">{item.author.displayName}</span>
+        <span className="flex shrink-0 items-center gap-1">
+          <Icon icon="tabler:download" className="size-3.5" />
+          {item.installCount}
+        </span>
+      </div>
+    </button>
   )
 }
