+++ src/entrypoints/options/pages/custom-actions/custom-actions-config.tsx
@@ -4,29 +4,34 @@ import { ConfigCard } from "../../components/config-card"
 import { EntityEditorLayout } from "../../components/entity-editor-layout"
 import { CustomActionConfigForm } from "./action-config-form"
 import { CustomActionCardList } from "./components/action-card-list"
+import { ActionCollectionProvider } from "./components/action-collection-detail-dialog"
+import { ActionCollectionInstallListener } from "./components/action-collection-install-listener"
 import { FeaturedActionCollections } from "./components/featured-collections"
 import { MyPublishedButton } from "./components/my-published-button"
 
 export function CustomActionsConfig() {
   return (
-    <ConfigCard
-      id="custom-actions"
-      title={(
-        <div className="flex w-full items-center justify-between gap-2">
-          <span className="inline-flex items-center gap-2">
-            {i18n.t("options.floatingButtonAndToolbar.selectionToolbar.customActions.title")}
-            <Badge variant="secondary" className="text-xs font-medium">Public Beta</Badge>
-          </span>
-          <MyPublishedButton />
+    <ActionCollectionProvider>
+      <ActionCollectionInstallListener />
+      <ConfigCard
+        id="custom-actions"
+        title={(
+          <div className="flex w-full items-center justify-between gap-2">
+            <span className="inline-flex items-center gap-2">
+              {i18n.t("options.floatingButtonAndToolbar.selectionToolbar.customActions.title")}
+              <Badge variant="secondary" className="text-xs font-medium">Public Beta</Badge>
+            </span>
+            <MyPublishedButton />
+          </div>
+        )}
+        description={i18n.t("options.floatingButtonAndToolbar.selectionToolbar.customActions.description")}
+        className="lg:flex-col"
+      >
+        <div className="flex flex-col gap-4">
+          <FeaturedActionCollections />
+          <EntityEditorLayout list={<CustomActionCardList />} editor={<CustomActionConfigForm />} />
         </div>
-      )}
-      description={i18n.t("options.floatingButtonAndToolbar.selectionToolbar.customActions.description")}
-      className="lg:flex-col"
-    >
-      <div className="flex flex-col gap-4">
-        <FeaturedActionCollections />
-        <EntityEditorLayout list={<CustomActionCardList />} editor={<CustomActionConfigForm />} />
-      </div>
-    </ConfigCard>
+      </ConfigCard>
+    </ActionCollectionProvider>
   )
 }
