+++ src/utils/constants/action-collection.ts
@@ -1,0 +1,1 @@
+export const INSTALL_ACTION_COLLECTION_PARAM = "installActionCollection"
